Bletcherous hack to get index page working again
There is something strange going on with pagination.  With the default
pages set to 10, I am now getting an index page that shows the list of
pages ("2", "3", etc), but nothing on that index page.  Clicking to 2,
3, etc works just fine.

This appears to have started once I added my 10th entry under
/content/italiano.  I'm not yet sure why this is going on, but setting
pagination to 20 restores that index page.  I assume anything more
than the number of italiano posts would do the trick.

Rough guess: this is something to do with either:

- the `range where Paginator.Pages "Type" "==" "posts"` in
m10c/layouts/index.html

- or the duplicate pagination under layouts/_default/list.html --
which is a problem; see https://gohugo.io/templates/pagination/.